moved all availableExercises to training service
