+++ src/app/training/new-training/new-training.component.ts
@@ -1,18 +1,19 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { AngularFirestore } from '@angular/fire/compat/firestore';
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { Subscription } from 'rxjs';
 
 import { TrainingService } from '../training.service';
+import { Exercise } from '../exercise.model';
 
 @Component({
   selector: 'app-new-training',
   templateUrl: './new-training.component.html',
   styleUrls: ['./new-training.component.css']
 })
-export class NewTrainingComponent implements OnInit {
-  exercises: Observable<any> | undefined;
+export class NewTrainingComponent implements OnInit, OnDestroy {
+  exercises: Exercise[] | undefined;
+  exerciseSubscription: Subscription | undefined;
 
   constructor(
     private trainingService: TrainingService,
@@ -20,29 +21,17 @@ export class NewTrainingComponent implements OnInit {
   ) { }
 
   ngOnInit(): void {
-    // this.exercises = this.db
-    //   .collection( 'availableExercises' )
-    //   .valueChanges();
-
-    this.exercises = this.db
-      .collection( 'availableExercises')
-      .snapshotChanges()
-      .pipe(
-        map( docData => {
-          return docData.map( doc => {
-            let object1 = doc.payload.doc.data()
-
-            return {
-              id: doc.payload.doc.id,
-              ...( typeof object1 === 'object' )? object1 : {}
-            };
-
-          });
-        })
-      )
+    this.exerciseSubscription = this.trainingService.exercisesChanged.subscribe( exercises => {
+      this.exercises = exercises
+    })
+    this.trainingService.fetchAvailableExercises()
   }
 
   onStartTraining( form: NgForm ) {
     this.trainingService.startExercise( form.value.exercise )
+  }
+
+  ngOnDestroy(): void {
+    this.exerciseSubscription?.unsubscribe();
   }
 }
